refactor(ui): move reusable hooks to src/hooks/
