+++ src/app/index.tsx
@@ -6,7 +6,7 @@ import { App } from "../ui/App"
 import { parseArgs, type ParsedArgs } from "./args"
 import { env } from "../env"
 import { loadSettings } from "../filestore"
-import { createNoodleKeymap } from "../ui/useKeymap"
+import { createNoodleKeymap } from "../hooks/useKeymap"
 import { parseOverrides } from "../ui/keybind"
 import { join } from "node:path"
 import * as yaml from "js-yaml"
