+++ src/hooks/useJsonHighlight.ts
@@ -1,8 +1,8 @@
 import { useCallback, useEffect, useRef } from "react"
 import { SyntaxStyle } from "@opentui/core"
 import type { TextareaRenderable, LineNumberRenderable } from "@opentui/core"
-import type { Theme } from "./theme-data"
-import { highlightJsonTokens } from "./syntax"
+import type { Theme } from "../ui/theme-data"
+import { highlightJsonTokens } from "../ui/syntax"
 import type { Environment } from "../schema"
 
 function createJsonSyntaxStyle(theme: Theme): SyntaxStyle {
