+++ src/ui/AppInner.tsx
@@ -5,13 +5,13 @@ import { Sidebar } from "./Sidebar"
 import { UrlBar } from "./UrlBar"
 import { RequestPane } from "./RequestPane"
 import { ResponsePane } from "./ResponsePane"
-import { useCollection } from "./useCollection"
-import { useSidebarSelection } from "./useSidebarSelection"
-import { useResponse } from "./useResponse"
-import { useRequestDraft } from "./useRequestDraft"
-import { useEditBrowse } from "./useEditBrowse"
-import { useEnvironments } from "./useEnvironments"
-import { useEnvironmentEditor } from "./useEnvironmentEditor"
+import { useCollection } from "../hooks/useCollection"
+import { useSidebarSelection } from "../hooks/useSidebarSelection"
+import { useResponse } from "../hooks/useResponse"
+import { useRequestDraft } from "../hooks/useRequestDraft"
+import { useEditBrowse } from "../hooks/useEditBrowse"
+import { useEnvironments } from "../hooks/useEnvironments"
+import { useEnvironmentEditor } from "../hooks/useEnvironmentEditor"
 import { type Focus } from "./focus"
 import { HelpOverlay } from "./HelpOverlay"
 import { ConfirmOverlay } from "./ConfirmOverlay"
