+++ src/ui/JsonBodyViewer.tsx
@@ -1,7 +1,7 @@
 import { useEffect, useRef } from "react"
 import type { TextareaRenderable } from "@opentui/core"
 import type { Theme } from "./theme-data"
-import { highlightTextarea } from "./useJsonHighlight"
+import { highlightTextarea } from "../hooks/useJsonHighlight"
 import type { Environment } from "../schema"
 
 export function JsonBodyViewer({
