+++ src/ui/index.ts
@@ -5,12 +5,12 @@ export { RequestPane } from "./RequestPane"
 export { ResponsePane } from "./ResponsePane"
 export { StatusBar, statusBarText } from "./StatusBar"
 export type { StatusBarSections } from "./StatusBar"
-export { useCollection } from "./useCollection"
-export type { UseCollectionResult } from "./useCollection"
-export { useSidebarSelection } from "./useSidebarSelection"
-export type { UseSidebarSelectionResult } from "./useSidebarSelection"
-export { useResponse } from "./useResponse"
-export type { UseResponseResult } from "./useResponse"
+export { useCollection } from "../hooks/useCollection"
+export type { UseCollectionResult } from "../hooks/useCollection"
+export { useSidebarSelection } from "../hooks/useSidebarSelection"
+export type { UseSidebarSelectionResult } from "../hooks/useSidebarSelection"
+export { useResponse } from "../hooks/useResponse"
+export type { UseResponseResult } from "../hooks/useResponse"
 export { nextIndex } from "./selection"
 export { startSend, finishSend, failSend } from "./sendState"
 export type { SendState } from "./sendState"
