+++ src/ui/useAppKeymap.ts
@@ -3,10 +3,10 @@ import { join } from "node:path"
 import type { MutableRefObject } from "react"
 import { cycleFocus, type Focus } from "./focus"
 import type { Keybinds } from "./keybind"
-import type { UseEditBrowseResult } from "./useEditBrowse"
-import type { UseRequestDraftResult } from "./useRequestDraft"
-import type { UseEnvironmentsResult } from "./useEnvironments"
-import type { UseEnvironmentEditorResult } from "./useEnvironmentEditor"
+import type { UseEditBrowseResult } from "../hooks/useEditBrowse"
+import type { UseRequestDraftResult } from "../hooks/useRequestDraft"
+import type { UseEnvironmentsResult } from "../hooks/useEnvironments"
+import type { UseEnvironmentEditorResult } from "../hooks/useEnvironmentEditor"
 import type { Collection } from "../schema"
 
 export interface UseAppKeymapRefs {
